syn-data/requirements.txt
openai>=1.0.0
pyyaml>=6.0
jsonschema>=4.0.0
requests>=2.25.0 
bespokelabs-curator